+++ programs/drift/src/instructions/admin.rs
@@ -1041,8 +1041,9 @@ pub fn handle_initialize_perp_market(
             last_oracle_valid: false,
             target_base_asset_amount_per_lp: 0,
             per_lp_base: 0,
-            padding1: 0,
-            padding2: 0,
+            oracle_slot_delay_override: 0,
+            taker_speed_bump_override: 0,
+            amm_spread_adjustment: 0,
             total_fee_earned_per_lp: 0,
             net_unsettled_funding_pnl: 0,
             quote_asset_amount_with_unsettled_lp: 0,
@@ -3911,7 +3912,7 @@ pub fn handle_update_perp_market_number_of_users(
 }
 
 pub fn handle_update_perp_market_fuel(
-    ctx: Context<AdminUpdatePerpMarketFuel>,
+    ctx: Context<HotAdminUpdatePerpMarket>,
     fuel_boost_taker: Option<u8>,
     fuel_boost_maker: Option<u8>,
     fuel_boost_position: Option<u8>,
@@ -3985,6 +3986,66 @@ pub fn handle_update_perp_market_protected_maker_params(
         msg!("perp_market.protected_maker_dynamic_divisor: unchanged");
     }
 
+    Ok(())
+}
+
+#[access_control(
+    perp_market_valid(&ctx.accounts.perp_market)
+)]
+pub fn handle_update_perp_market_taker_speed_bump_override(
+    ctx: Context<HotAdminUpdatePerpMarket>,
+    taker_speed_bump_override: i8,
+) -> Result<()> {
+    let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
+    msg!(
+        "perp_market.amm.taker_speed_bump_override: {:?} -> {:?}",
+        perp_market.amm.taker_speed_bump_override,
+        taker_speed_bump_override
+    );
+
+    perp_market.amm.taker_speed_bump_override = taker_speed_bump_override;
+    Ok(())
+}
+
+#[access_control(
+    perp_market_valid(&ctx.accounts.perp_market)
+)]
+pub fn handle_update_perp_market_amm_spread_adjustment(
+    ctx: Context<HotAdminUpdatePerpMarket>,
+    amm_spread_adjustment: i8,
+) -> Result<()> {
+    let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
+    msg!(
+        "perp_market.amm.amm_spread_adjustment: {:?} -> {:?}",
+        perp_market.amm.amm_spread_adjustment,
+        amm_spread_adjustment
+    );
+
+    perp_market.amm.amm_spread_adjustment = amm_spread_adjustment;
+    Ok(())
+}
+
+#[access_control(
+    perp_market_valid(&ctx.accounts.perp_market)
+)]
+pub fn handle_update_perp_market_oracle_slot_delay_override(
+    ctx: Context<HotAdminUpdatePerpMarket>,
+    oracle_slot_delay_override: i8,
+) -> Result<()> {
+    let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
+    msg!(
+        "perp_market.amm.oracle_slot_delay_override: {:?} -> {:?}",
+        perp_market.amm.oracle_slot_delay_override,
+        oracle_slot_delay_override
+    );
+
+    perp_market.amm.oracle_slot_delay_override = oracle_slot_delay_override;
     Ok(())
 }
 
@@ -4908,7 +4969,7 @@ pub struct AdminUpdatePerpMarket<'info> {
 }
 
 #[derive(Accounts)]
-pub struct AdminUpdatePerpMarketFuel<'info> {
+pub struct HotAdminUpdatePerpMarket<'info> {
     #[account(
         constraint = admin.key() == admin_hot_wallet::id() || admin.key() == state.admin
     )]
@@ -5100,10 +5161,10 @@ pub struct AdminUpdatePerpMarketOracle<'info> {
 
 #[derive(Accounts)]
 pub struct AdminDisableBidAskTwapUpdate<'info> {
-    pub admin: Signer<'info>,
     #[account(
-        has_one = admin
+        constraint = admin.key() == admin_hot_wallet::id() || admin.key() == state.admin
     )]
+    pub admin: Signer<'info>,
     pub state: Box<Account<'info, State>>,
     #[account(mut)]
     pub user_stats: AccountLoader<'info, UserStats>,