+++ sdk/src/driftClient.ts
@@ -274,7 +274,8 @@ export class DriftClient {
 		this.program = new Program(
 			driftIDL as Idl,
 			config.programID ?? new PublicKey(DRIFT_PROGRAM_ID),
-			this.provider
+			this.provider,
+			config.coder
 		);
 
 		this.authority = config.authority ?? this.wallet.publicKey;
@@ -2540,7 +2541,8 @@ export class DriftClient {
 		associatedTokenAccount: PublicKey,
 		subAccountId?: number,
 		reduceOnly = false,
-		txParams?: TxParams
+		txParams?: TxParams,
+		initSwiftAccount = false
 	): Promise<VersionedTransaction | Transaction> {
 		const instructions = await this.getDepositTxnIx(
 			amount,
@@ -2550,7 +2552,24 @@ export class DriftClient {
 			reduceOnly
 		);
 
-		txParams = { ...(txParams ?? this.txParams), computeUnits: 600_000 };
+		if (initSwiftAccount) {
+			const isSignedMsgUserOrdersAccountInitialized =
+				await this.isSignedMsgUserOrdersAccountInitialized(
+					this.wallet.publicKey
+				);
+
+			if (!isSignedMsgUserOrdersAccountInitialized) {
+				const [, initializeSignedMsgUserOrdersAccountIx] =
+					await this.getInitializeSignedMsgUserOrdersAccountIx(
+						this.wallet.publicKey,
+						8
+					);
+
+				instructions.push(initializeSignedMsgUserOrdersAccountIx);
+			}
+		}
+
+		txParams = { ...(txParams ?? this.txParams), computeUnits: 800_000 };
 
 		const tx = await this.buildTransaction(instructions, txParams);
 
@@ -2572,7 +2591,8 @@ export class DriftClient {
 		associatedTokenAccount: PublicKey,
 		subAccountId?: number,
 		reduceOnly = false,
-		txParams?: TxParams
+		txParams?: TxParams,
+		initSwiftAccount = false
 	): Promise<TransactionSignature> {
 		const tx = await this.createDepositTxn(
 			amount,
@@ -2580,7 +2600,8 @@ export class DriftClient {
 			associatedTokenAccount,
 			subAccountId,
 			reduceOnly,
-			txParams
+			txParams,
+			initSwiftAccount
 		);
 
 		const { txSig, slot } = await this.sendTransaction(tx, [], this.opts);
@@ -8966,7 +8987,9 @@ export class DriftClient {
 		enteringHighLeverageMode?: boolean
 	) {
 		let feeTier;
-		if (user) {
+		const userHLM =
+			(user?.isHighLeverageMode() ?? false) || enteringHighLeverageMode;
+		if (user && !userHLM) {
 			feeTier = user.getUserFeeTier(marketType);
 		} else {
 			const state = this.getStateAccount();
@@ -8988,7 +9011,7 @@ export class DriftClient {
 			}
 
 			takerFee += (takerFee * marketAccount.feeAdjustment) / 100;
-			if (user && (user.isHighLeverageMode() || enteringHighLeverageMode)) {
+			if (userHLM) {
 				takerFee *= 2;
 			}
 			makerFee += (makerFee * marketAccount.feeAdjustment) / 100;