+++ sdk/src/types.ts
@@ -625,6 +625,10 @@ export type OrderActionRecord = {
 	makerOrderCumulativeQuoteAssetAmountFilled: BN | null;
 	oraclePrice: BN;
 	bitFlags: number;
+	takerExistingQuoteEntryAmount: BN | null;
+	takerExistingBaseAssetAmount: BN | null;
+	makerExistingQuoteEntryAmount: BN | null;
+	makerExistingBaseAssetAmount: BN | null;
 };
 
 export type SwapRecord = {
@@ -1001,6 +1005,9 @@ export type AMM = {
 	netUnsettledFundingPnl: BN;
 	quoteAssetAmountWithUnsettledLp: BN;
 	referencePriceOffset: number;
+
+	takerSpeedBumpOverride: number;
+	ammSpreadAdjustment: number;
 };
 
 // # User Account Types