+++ CHANGELOG.md
@@ -9,7 +9,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: perp market amm oracle delay override ([#1679](https://github.com/drift-labs/protocol-v2/pull/1679))
+- program: sanitize long tail perp market orders less frequently ([#1641](https://github.com/drift-labs/protocol-v2/pull/1641))
+
 ### Fixes
+
+### Breaking
+
+## [2.123.0] - 2025-06-13
+
+### Features
+
+- program: simplify user can skip duration ([#1668](https://github.com/drift-labs/protocol-v2/pull/1668))
+- program: allow limit orders without auctions in swift ([#1661](https://github.com/drift-labs/protocol-v2/pull/1661))
+- program: add taker_speed_bump_override and amm_spread_adjustment ([#1665](https://github.com/drift-labs/protocol-v2/pull/1665))
+
+### Fixes
+
+### Breaking
+
+## [2.122.0] - 2025-06-05
+
+### Features
+
+- program: add existing position fields to order records ([#1614](https://github.com/drift-labs/protocol-v2/pull/1614))
+
+### Fixes
+
+- sdk: fix to getMaxTradeSizeUSDCForPerp which was previously overshooting max allowed size due to IMF factor
+- program: check limit price after applying buffer in trigger limit order ([#1648](https://github.com/drift-labs/protocol-v2/pull/1648))
+- program: check limit price when setting auction for limit order ([#1650](https://github.com/drift-labs/protocol-v2/pull/1650))
 
 ### Breaking